# api/swagger.yaml
swagger: '2.0'
info:
  description: ''
  version: 1.0.0
  title: 挣闲钱
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: petstore.swagger.io
basePath: /
...
schemes:
  - https
  - http
paths:
...
  /task/qAcceptPage:
    get:
      tags:
        - task
      summary: query accepted tasks
      description: query a page of tasks accepted(获取一页用户接受的任务)
      operationId: qAcceptPage
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          description: >-
            if page is greater than 0, response a page of tasks(10), if page is
            0, response 3 front
            tasks(如果page参数为0，则返回3个任务，供导航页使用；如果page大于0，则按页数返回10个任务)
          required: true
          type: integer
        - name: userId
          in: query
          description: the id of user
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/tasks'
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
  /task/qPublishPage:
    get:
      tags:
        - task
      summary: query published tasks
      description: query a page of tasks published(获取一页用户发布的任务)
      operationId: qPublishPage
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          description: >-
            if page is greater than 0, response a page of tasks(10), if page is
            0, response 3 front
            tasks(如果page参数为0，则返回3个任务，供导航页使用；如果page大于0，则按页数返回10个任务)
          required: true
          type: integer
        - name: userId
          in: query
          description: the id of user
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/tasks'
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
  /task/queryPageQ:
    get:
      tags:
        - task
      summary: query questionare tasks
      description: >-
        query a page of questionareTasks not accepted and not published by
        user(获取一页用户没接受过，也不是由用户发布的问卷任务)
      operationId: queryPageQ
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          description: >-
            if page is greater than 0, response a page of tasks(10), if page is
            0, response 5 front
            tasks(如果page参数为0，则返回5个任务，供导航页使用；如果page大于0，则按页数返回10个任务)
          required: true
          type: integer
        - name: userId
          in: query
          description: the id of user
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/tasks'
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
  /task/queryPageD:
    get:
      tags:
        - task
      summary: query delivery tasks
      description: >-
        query a page of delivery tasks not accepted and not published by
        user(获取一页用户没接受过，也不是由用户发布的快递任务)
      operationId: queryPageD
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          description: >-
            if page is greater than 0, response a page of tasks(10), if page is
            0, response 5 front
            tasks(如果page参数为0，则返回5个任务，供导航页使用；如果page大于0，则按页数返回10个任务)
          required: true
          type: integer
        - name: userId
          in: query
          description: the id of user
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/tasks'
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
  /task/queryTitle:
    get:
      tags:
        - task
      summary: query tasks by title
      description: >-
        query some tasks not accepted and not published by user by
        title(通过标题查询获取用户没接受过，也不是由用户发布的任务)
      operationId: queryTitle
      produces:
        - application/json
      parameters:
        - name: title
          in: query
          description: the title of tasks
          required: true
          type: integer
        - name: userId
          in: query
          description: the id of user
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/tasks'
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
  '/task/questionare/{taskId}':
    get:
      tags:
        - task
      summary: get a questionareTask
      description: get a questionareTask by taskId(通过任务id获取一个问卷任务)
      operationId: getQuestionare
      produces:
        - application/json
      parameters:
        - name: taskId
          in: path
          description: the id of tasks
          required: true
          type: integer
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/questionareTask'
        '404':
          description: not found
          schema:
            type: object
            properties:
              error:
                type: string
...
definitions:
  task:
    type: object
    properties:
      taskId:
        type: integer
      taskType:
        type: string
      taskTitle:
        type: string
      endTime:
        type: string
      userId:
        type: integer
  questionare:
    type: object
    properties:
      queryid:
        type: integer
      content:
        type: string
...
  tasks:
    properties:
      contents:
        type: array
        items:
          $ref: '#/definitions/task'
  questionareTask:
    properties:
      task:
        $ref: '#/definitions/task'
      questionare:
        $ref: '#/definitions/questionare'
...